+++ pypeit/spectrographs/__init__.py
@@ -21,6 +21,8 @@ from pypeit.spectrographs import vlt_fors
 from pypeit.spectrographs import mdm_osmos
 from pypeit.spectrographs import lbt_luci
 from pypeit.spectrographs import mmt_binospec
+from pypeit.spectrographs import mmt_mmirs
+from pypeit.spectrographs import mmt_bluechannel
 from pypeit.spectrographs import not_alfosc
 from pypeit.spectrographs import p200_dbsp
 from pypeit.spectrographs import p200_tspec