+++ CHANGES.rst
@@ -1,4 +1,4 @@
-
+=======
 1.2.1dev
 --------
 
@@ -45,6 +45,7 @@
   alignment boxes from slitmask design.
 - Added a table with DEIMOS slitmask design and objects info to the
   SlitTraceSet datamodel
+- Add support for MMTO Blue Channel Spectrograph
 - Add GitHub Actions CI workflow
 - Incorporates a procedure to enable GMOS Nod and Shuffle observations
 - New GMOS wavelength solutions